refactor: controller, repository 최상위 point 도메인 파일로 이동
- 테스트 파일이 사라져서 폴더로 묶여있을 이유가 사라짐
