+++ src/point/point.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { PointController } from './controller/point.controller';
+import { PointController } from './point.controller';
 import { PointService } from './service/point.service';
 import { DatabaseModule } from 'src/database/database.module';
 import { PointPolicy } from './service/policy/point.policy';
-import { PointRepository } from './repository/point.repository';
+import { PointRepository } from './point.repository';
 
 @Module({
   imports: [DatabaseModule],
